Re-added Search task
